requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-genai>=1.0.0
python-dotenv>=1.0.0
markdownify>=0.11.0
aiohttp>=3.9.0
waybackpy>=3.0.0
